+++ Observer/EmailObserver.php
@@ -4,16 +4,19 @@ namespace Transbank\Webpay\Observer;
 
 use Magento\Framework\Event\ObserverInterface;
 
-class SendMail implements ObserverInterface
+class EmailObserver implements ObserverInterface
 {
 
     protected $orderSender;
+    protected $invoiceSender;
     protected $_logger;
+
 
     public function __construct (
         \Psr\Log\LoggerInterface $logger,
         \Magento\Sales\Model\Order $order,
         \Magento\Sales\Model\Order\Email\Sender\OrderSender $orderSender,
+        \Magento\Sales\Model\Order\Email\Sender\InvoiceSender $invoiceSender,
         \Magento\Sales\Model\Service\InvoiceService $invoiceService,
         \Magento\Framework\DB\Transaction $transaction,
         \Transbank\Webpay\Model\Config\ConfigProvider $configProvider)
@@ -21,6 +24,7 @@ class SendMail implements ObserverInterface
         $this->_logger = $logger;
         $this->order = $order;
         $this->orderSender = $orderSender;
+        $this->invoiceSender = $invoiceSender;
         $this->invoiceService = $invoiceService;
         $this->transaction = $transaction;
         $this->configProvider = $configProvider;
@@ -29,7 +33,6 @@ class SendMail implements ObserverInterface
     public function execute(\Magento\Framework\Event\Observer $observer) {
 
         $emailSettings = $this->configProvider->getEmailSettings();
-        $this->_logger->debug($emailSettings);
 
         $order = $observer->getEvent()->getOrder();
 
@@ -46,26 +49,6 @@ class SendMail implements ObserverInterface
             }
         }
 
-        $invoiceSettings = $this->configProvider->getInvoiceSettings();
-        if ($invoiceSettings == 'transbank') {
-
-            $order->addStatusHistoryComment('Automatically Invoiced by Transbank', true);
-
-            $this->_logger->debug('Creating Invoice');
-
-            $order->setCanSendNewEmailFlag(true);
-            $order->save();
-
-            if ($order->canInvoice()) {
-
-                $invoice = $this->invoiceService->prepareInvoice($order);
-                $invoice->register();
-                $invoice->save();
-                
-                $transactionSave = $this->transaction->addObject($invoice)->addObject($invoice->getOrder());
-                $transactionSave->save();
-            }
-        }
     }
